As far as possible, restrict the constructors for all MarshalObjectByRef types.
The only types which can be used outside of Eraser.Manager are the *Executor types, which is used for hosting, as well as the Task type, which needs to be created for use with the Executors. Since we need to ensure proxies are properly used when needed, ensure that the Task type is constructed only after an Executor type is constructed, allowing Executor types to inject remoting calls.